Add example on how to use IdlingResources as an extra unit.
The Countdown is using a Thread to run instead of any other thing only for demo purposes. Please, don't ever use plain Threads in Android. This is just to showcase that Espresso doesn't monitor Threads and how you can implement an IdlingResource in your test in order to wait for some resource that Espresso is not aware of. For more examples look at the Espresso source code (IdlingResource, CountingIdlingResource or OnDemandIdlingResource)